+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- [25-1-25-1] Fix:ratewindow failed to change stats when contest changed.
 - [25-1-25] Fix app not following radio mode. Fix crash on recalc mults.
 - [25-1-24] Fix mismatched quotes causing crash.
 - [25-1-23] Added a basic rate window. Changed DB to WAL mode.